+++ model/model.go
@@ -2,7 +2,6 @@
 package model
 
 import (
-	"github.com/zeroibot/pack/ds"
 	"github.com/zeroibot/pack/dyn"
 	"github.com/zeroibot/pack/qb"
 )
@@ -16,10 +15,10 @@ type Schema[T any] struct {
 }
 
 // NewSchema creates a new Schema and registers its type to qb
-func NewSchema[T any](this *qb.Instance, structRef *T, table string) ds.Result[*Schema[T]] {
+func NewSchema[T any](this *qb.Instance, structRef *T, table string) (*Schema[T], error) {
 	err := qb.AddType(this, structRef)
 	if err != nil {
-		return ds.Error[*Schema[T]](err)
+		return nil, err
 	}
 	schema := new(Schema[T]{
 		Name:     dyn.TypeName(structRef),
@@ -28,28 +27,28 @@ func NewSchema[T any](this *qb.Instance, structRef *T, table string) ds.Result[*
 		Reader:   qb.FullRowReader(this, structRef),
 		Instance: this,
 	})
-	return ds.NewResult(schema, nil)
+	return schema, nil
 }
 
 // NewSharedSchema creates a new shared Schema (no table)
-func NewSharedSchema[T any](this *qb.Instance, structRef *T) ds.Result[*Schema[T]] {
+func NewSharedSchema[T any](this *qb.Instance, structRef *T) (*Schema[T], error) {
 	return NewSchema(this, structRef, "")
 }
 
 // AddSchema adds a new Schema and adds error to the error list if any
 func AddSchema[T any](this *qb.Instance, structRef *T, table string, errs []error) *Schema[T] {
-	result := NewSchema(this, structRef, table)
-	if result.IsError() {
-		errs = append(errs, result.Error())
+	schema, err := NewSchema(this, structRef, table)
+	if err != nil {
+		errs = append(errs, err)
 	}
-	return result.Value()
+	return schema
 }
 
 // AddSharedSchema adds a new shared Schema (no table) and adds error to the error list if any
 func AddSharedSchema[T any](this *qb.Instance, structRef *T, errs []error) *Schema[T] {
-	result := NewSharedSchema(this, structRef)
-	if result.IsError() {
-		errs = append(errs, result.Error())
+	schema, err := NewSharedSchema(this, structRef)
+	if err != nil {
+		errs = append(errs, err)
 	}
-	return result.Value()
+	return schema
 }