+++ model/schema_group.go
@@ -1,57 +1,58 @@
 package model
 
 import (
-	"github.com/zeroibot/pack/ds"
 	"github.com/zeroibot/pack/fail"
 	"github.com/zeroibot/pack/my"
 	"github.com/zeroibot/pack/qb"
 )
 
 // Count performs a CountQuery at schema table
-func (s *Schema[T]) Count(rq *my.Request, condition qb.DualCondition[T]) ds.Result[int] {
+func (s *Schema[T]) Count(rq *my.Request, condition qb.DualCondition[T]) (int, error) {
 	return s.countAt(rq, condition, s.Table)
 }
 
 // CountAt performs a CountQuery at the given table
-func (s *Schema[T]) CountAt(rq *my.Request, condition qb.DualCondition[T], table string) ds.Result[int] {
+func (s *Schema[T]) CountAt(rq *my.Request, condition qb.DualCondition[T], table string) (int, error) {
 	return s.countAt(rq, condition, table)
 }
 
 // Sum performs a SumQuery at schema table
-func (s *Schema[T]) Sum(rq *my.Request, columns []string, reader qb.RowReader[T], condition qb.DualCondition[T]) ds.Result[T] {
+func (s *Schema[T]) Sum(rq *my.Request, columns []string, reader qb.RowReader[T], condition qb.DualCondition[T]) (T, error) {
 	return s.sumAt(rq, columns, reader, condition, s.Table)
 }
 
 // SumAt performs a SumQuery at the given table
-func (s *Schema[T]) SumAt(rq *my.Request, columns []string, reader qb.RowReader[T], condition qb.DualCondition[T], table string) ds.Result[T] {
+func (s *Schema[T]) SumAt(rq *my.Request, columns []string, reader qb.RowReader[T], condition qb.DualCondition[T], table string) (T, error) {
 	return s.sumAt(rq, columns, reader, condition, table)
 }
 
 // Common: create and execute CountQuery at given table
-func (s *Schema[T]) countAt(rq *my.Request, condition qb.DualCondition[T], table string) ds.Result[int] {
+func (s *Schema[T]) countAt(rq *my.Request, condition qb.DualCondition[T], table string) (int, error) {
 	// Check that condition is set
 	if condition == nil {
 		rq.Fail(my.Err500, "Count condition is not set")
-		return ds.Error[int](fail.MissingParams)
+		return 0, fail.MissingParams
 	}
 
 	// Build CountQuery and execute
 	q := qb.NewCountQuery[T](s.Instance, table)
 	q.Where(condition)
 
-	result := q.Count(rq.DB)
-	if result.IsError() {
+	count, err := q.Count(rq.DB)
+	if err != nil {
 		rq.Status = my.Err500
+		return 0, err
 	}
-	return result
+	return count, nil
 }
 
 // Common: create and execute SumQuery at given table
-func (s *Schema[T]) sumAt(rq *my.Request, columns []string, reader qb.RowReader[T], condition qb.DualCondition[T], table string) ds.Result[T] {
+func (s *Schema[T]) sumAt(rq *my.Request, columns []string, reader qb.RowReader[T], condition qb.DualCondition[T], table string) (T, error) {
+	var zero T
 	// Check that condition is set
 	if condition == nil {
 		rq.Fail(my.Err500, "Sum condition is not set")
-		return ds.Error[T](fail.MissingParams)
+		return zero, fail.MissingParams
 	}
 
 	// Build SumQuery and execute
@@ -60,9 +61,10 @@ func (s *Schema[T]) sumAt(rq *my.Request, columns []string, reader qb.RowReader[
 	q.Columns(this, columns...)
 	q.Where(condition)
 
-	result := q.Sum(rq.DB)
-	if result.IsError() {
+	result, err := q.Sum(rq.DB)
+	if err != nil {
 		rq.Status = my.Err500
+		return zero, err
 	}
-	return result
+	return result, nil
 }